Create LEDC.sol
This smart contract is an optimized version of an ERC20 token contract, which is gas-efficient and fast. It uses OpenZeppelin's SafeMath library to perform arithmetic operations in a safe way, avoiding potential overflow and underflow errors. It also has the transferFrom and approve functions which allows for token transfers by another address on behalf of the token holder and also the allowance function which allows a token holder to check the allowance of another address.

This contract also allows for token exchange for other crypto, and it allows users to interact and share their health data with hospitals, doctors, medical care centers, coaches, etc. 

We are developing smart contracts and decentralized applications that allow users to interact with their tokens and share their health data with various stakeholders.